+++ kiosk-demo-atablefor/script/book_flow.rb
@@ -18,6 +18,7 @@ require "date"
 require "jwt"
 require "json"
 require "net/http"
+require "kiosk/redteam/wire"
 require "uri"
 require "openssl"
 require "securerandom"
@@ -29,7 +30,7 @@ def post_json(url, body, headers = {})
   uri = URI(url)
   req = Net::HTTP::Post.new(uri, { "Content-Type" => "application/json" }.merge(headers))
   req.body = JSON.generate(body)
-  res = Net::HTTP.new(uri.host, uri.port).request(req)
+  res = Kiosk::Redteam::Wire.http_for(uri).request(req)
   [res.code.to_i, (JSON.parse(res.body) rescue {})]
 end
 
@@ -42,7 +43,7 @@ end
 def get_json(url, params = {}, headers = {})
   uri = URI(url)
   uri.query = URI.encode_www_form(params) unless params.empty?
-  res = Net::HTTP.new(uri.host, uri.port).request(Net::HTTP::Get.new(uri, headers))
+  res = Kiosk::Redteam::Wire.http_for(uri).request(Net::HTTP::Get.new(uri, headers))
   [res.code.to_i, (JSON.parse(res.body) rescue {})]
 end